telegram_bot/requirements.txt
python-telegram-bot>=20.0
vosk==0.3.45
